feat: float to double

+++ src/image.rs
@@ -7,7 +7,7 @@ pub fn export_ppm_image(width: u32, height: u32) {
     for y in 0..height {
         // eprintln!("Scan-lines processed: {}/{}", y, height);
         for x in 0..width {
-            let pixel_color = Color3::new(x as f32, (255 - y) as f32, 67 as f32);
+            let pixel_color = Color3::new(x as f64, (255 - y) as f64, 67 as f64);
 
             // println!("{} {} {}", pixel_color.x, pixel_color.y, pixel_color.z);
         }
